--- src/scientio/concept.py
@@ -1,11 +1,0 @@
-# I'm a stub
-from enum import Enum
-
-class Concept(object):
-    def __init__(self):
-        self.properties = set()
-        self.relationships = set()
-
-
-
-
